sf_police_requests/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.1.2
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
flake8==5.0.4
mccabe==0.7.0
packaging==21.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
sqlparse==0.4.3
wrapt==1.14.1